refactor: Admin Test도 동일하게 수정
